Refactor Eventing Components
