+++ components/event-bus/README.md
@@ -8,14 +8,14 @@ The Event Bus enables Kyma to integrate with various other external solutions. T
 
 Currently, Event Bus makes the following three Docker images available to the `kyma core` Helm chart:
 
-- event-bus-publish-knative
-- event-bus-subscription-controller-knative
+- event-publish-service
+- subscription-controller
 
 There are also end-to-end test Docker images to use as `helm tests`. See [the tests in the `event-bus` directory](https://github.com/kyma-project/kyma/tree/master/tests/event-bus) for more details.
 
 ## Development
 
-The two binaries of `Event Bus` reside under `cmd/event-bus-XXXX` "e.g. `cmd/event-bus-publish-knative`". They each have a Makefile to build and test the component as well as to create and push a Docker image. The following table explains the various make targets.
+The two binaries of `Event Bus` reside under `cmd/event-bus-XXXX` "e.g. `cmd/event-publish-service`". They each have a Makefile to build and test the component as well as to create and push a Docker image. The following table explains the various make targets.
 
 
 |Command| Description|
