+++ components/event-bus/before-commit.sh
@@ -33,8 +33,8 @@ if [ "$1" == "$CI_FLAG" ]; then
 	buildEnv="env CGO_ENABLED=0"
 fi
 
-${buildEnv} go build -v -a -installsuffix cgo -o ./cmd/event-bus-publish-knative/event-bus-publish-knative ./cmd/event-bus-publish-knative/
-${buildEnv} go build -v -a -installsuffix cgo -o ./cmd/event-bus-subscription-controller-knative/event-bus-subscription-controller-knative ./cmd/event-bus-subscription-controller-knative/...
+${buildEnv} go build -v -a -installsuffix cgo -o ./cmd/event-publish-service/event-publish-service ./cmd/event-publish-service/
+${buildEnv} go build -v -a -installsuffix cgo -o ./cmd/subscription-controller/subscription-controller ./cmd/subscription-controller/...
 
 goBuildResult=$?
 if [ ${goBuildResult} != 0 ]; then
@@ -120,5 +120,6 @@ for vPackage in "${packagesToVet[@]}"; do
 	fi
 done
 
-rm ./cmd/event-bus-publish-knative/event-bus-publish-knative
-rm ./cmd/event-bus-subscription-controller-knative/event-bus-subscription-controller-knative
+rm ./cmd/event-publish-service/event-publish-service
+rm ./cmd/subscription-controller/subscription-controller
+
