+++ components/event-bus/cmd/event-publish-service/Dockerfile
@@ -9,7 +9,7 @@ COPY internal/knative ./internal/knative
 COPY internal/ea      ./internal/ea
 COPY licenses         ./licenses
 
-WORKDIR /go/src/github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-publish-knative/
+WORKDIR /go/src/github.com/kyma-project/kyma/components/event-bus/cmd/event-publish-service/
 COPY main.go     .
 COPY application ./application
 COPY handlers    ./handlers
@@ -19,7 +19,7 @@ COPY publisher   ./publisher
 COPY validators  ./validators
 COPY util        ./util
 
-RUN CGO_ENABLED=0 GOOS=linux go build -v -a -installsuffix cgo -o event-bus-publish-knative .
+RUN CGO_ENABLED=0 GOOS=linux go build -v -a -installsuffix cgo -o event-publish-service .
 
 FROM alpine:3.7
 LABEL source=git@github.com:kyma-project/kyma.git
@@ -30,9 +30,9 @@ ENV APP_VERSION $version
 WORKDIR /root/
 RUN apk --no-cache upgrade && apk --no-cache add curl
 
-COPY --from=builder /go/src/github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-publish-knative/event-bus-publish-knative .
+COPY --from=builder /go/src/github.com/kyma-project/kyma/components/event-bus/cmd/event-publish-service/event-publish-service .
 COPY --from=builder /go/src/github.com/kyma-project/kyma/components/event-bus/licenses ./licenses
 
 EXPOSE 8080
 
-ENTRYPOINT ["/root/event-bus-publish-knative"]
+ENTRYPOINT ["/root/event-publish-service"]
