+++ components/event-bus/cmd/event-publish-service/Makefile
@@ -1,4 +1,4 @@
-NAME = kyma-project/event-bus-publish-knative
+NAME = kyma-project/event-publish-service
 VERSION = 0.1.0
 COMPONENT = event-bus
 REGISTRY = eu.gcr.io
@@ -12,7 +12,7 @@ clean:
 	rm -rf docker/
 
 compile:
-	go build -o bin/event-bus-publish-knative
+	go build -o bin/event-publish-service
 
 docker-build:
 	./dockerBuild.sh  $(NAME) $(VERSION) $(COMPONENT)
