+++ components/event-bus/cmd/event-publish-service/dockerBuild.sh
