+++ components/event-bus/cmd/event-publish-service/handlers/publish.go
@@ -12,9 +12,9 @@ import (
 
 	"github.com/gofrs/uuid"
 	api "github.com/kyma-project/kyma/components/event-bus/api/publish"
-	"github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-publish-knative/publisher"
-	constants "github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-publish-knative/util"
-	"github.com/kyma-project/kyma/components/event-bus/cmd/event-bus-publish-knative/validators"
+	"github.com/kyma-project/kyma/components/event-bus/cmd/event-publish-service/publisher"
+	constants "github.com/kyma-project/kyma/components/event-bus/cmd/event-publish-service/util"
+	"github.com/kyma-project/kyma/components/event-bus/cmd/event-publish-service/validators"
 	"github.com/kyma-project/kyma/components/event-bus/internal/knative/publish/opts"
 	knative "github.com/kyma-project/kyma/components/event-bus/internal/knative/util"
 	"github.com/kyma-project/kyma/components/event-bus/internal/trace"
