+++ components/event-bus/cmd/event-publish-service/handlers/readiness_check.go
