+++ components/event-bus/cmd/event-publish-service/validators/publish_validator.go
