+++ components/event-bus/cmd/subscription-controller/cmd/manager/main.go
