+++ components/event-bus/cmd/subscription-controller/dockerBuild.sh
